+++ geodepoly/__init__.py
@@ -8,6 +8,8 @@ from .hyper_catalan import (
 )
 from .formal import FormalSeries
 from .series import series_root
+from .poly import Polynomial
+from .numeric import newton as newton_solve, aberth as aberth_solve, dk as dk_solve, companion_roots
 from .batched import (
     batched_poly_eval,
     batched_newton_step,
@@ -29,4 +31,9 @@ __all__ = [
     "torch_root_layer",
     "FormalSeries",
     "series_root",
+    "Polynomial",
+    "newton_solve",
+    "aberth_solve",
+    "dk_solve",
+    "companion_roots",
 ]